Add tests for onSelect prop
